refactor(flag_mapping): Split 5,215-line file into modular structure
Refactor flag_mapping.rs into a directory module with separate files for
better organization and maintainability:

- types.rs: FlagType, FlagMapping, CheckStatus, Accessibility, etc.
- scenes.rs: Scene ID constants (DEKU_TREE, FOREST_TEMPLE, etc.)
- mappings.rs: Static mapping tables (~3,000 lines of add_mapping calls)
- queries.rs: Lookup functions (get_mapping, get_mapped_locations, etc.)
- mq.rs: Master Quest integration functions
- eval.rs: OotEvalContext for logic evaluation
- checking.rs: Location status checking functions
- tests.rs: All unit tests
- mod.rs: Re-exports maintaining backward compatibility

All existing functionality preserved. All 60 tests pass.

Resolves #639
