+++ crate/oottracker/src/flag_mapping/mappings.rs
@@ -1,386 +1,23 @@
-//! OoT Flag Data Structures and Location Mappings
+//! Static mapping tables for OoT locations.
 //!
-//! This module provides the foundation for detecting location checks by mapping
-//! location IDs to memory flag addresses. It defines flag types, scene IDs, and
-//! stub mappings for all OoT locations imported from OoTMM world data.
-//!
-//! # Flag Structure (based on OoTMM research)
-//!
-//! OoT save data contains several types of flags stored in different memory regions:
-//!
-//! ## Scene Flags (per-scene, stored in save context at 0x00D4)
-//!
-//! Each scene has a 0x1C byte entry containing:
-//! - `chests` (u32 at offset 0x00): Opened chest flags
-//! - `switches` (u32 at offset 0x04): Switch/trigger flags
-//! - `room_clear` (u32 at offset 0x08): Room clear flags
-//! - `collectible` (u32 at offset 0x0C): Collectible item flags
-//! - `unused` (u32 at offset 0x10): Unused flags (sometimes repurposed)
-//! - `visited_rooms` (u32 at offset 0x14): Room visited flags
-//! - `visited_floors` (u32 at offset 0x18): Floor visited flags
-//!
-//! ## Global Flags
-//!
-//! - Gold Skulltulas: Stored in save context at 0x0E9C (0x18 bytes)
-//! - Event flags (event_chk_inf): Global event tracking at 0x0ED4 (0x1C bytes)
-//! - Item get flags (item_get_inf): Item obtained tracking at 0x0EF0 (0x08 bytes)
-//! - Info table (inf_table): NPC/misc flags at 0x0EF8 (0x3C bytes)
-//!
-//! # Usage
-//!
-//! ```ignore
-//! use oottracker::flag_mapping::{FlagMapping, get_mapping, get_all_oot_mappings};
-//!
-//! // Get mapping for a specific location
-//! if let Some(mapping) = get_mapping("oot_deku_tree_compass_chest") {
-//!     println!("Location {} is in scene {:?}", mapping.location_id, mapping.scene_id);
-//! }
-//!
-//! // Get all OoT location mappings
-//! for mapping in get_all_oot_mappings() {
-//!     println!("{}: {:?}", mapping.location_id, mapping.flag_type);
-//! }
-//! ```
+//! This module contains the large static data tables mapping location IDs
+//! to their flag addresses. The data is lazily initialized at runtime.
 
 use std::collections::HashMap;
 
 use once_cell::sync::Lazy;
-
 use ootmm::embedded_data;
-use ootmm::expr::{eval_str, EvalContext};
-use ootmm::is_region_reachable;
 use ootmm::region::Game;
-use ootmm::settings::RandomizerSettings;
-use ootmm::world_database::WorldDatabase;
-use ootr::{
-    model::{Dungeon, MainDungeon},
-    region::Mq,
-};
 
-use crate::world_database;
+use super::scenes as scene;
+use super::types::{FlagMapping, FlagType};
 
-// ============================================================================
-// Flag Type Definition
-// ============================================================================
-
-/// Types of flags used to track location checks in OoT save data.
-///
-/// Each location in the game is tracked by one of these flag types,
-/// stored in specific memory regions within the save context.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum FlagType {
-    /// Chest opened flags (scene flags offset 0x00).
-    /// Each bit represents a chest in the scene.
-    Chest,
-
-    /// Switch/trigger flags (scene flags offset 0x04).
-    /// Includes crystal switches, floor switches, etc.
-    Switch,
-
-    /// Room clear flags (scene flags offset 0x08).
-    /// Set when all enemies in a room are defeated.
-    RoomClear,
-
-    /// Collectible item flags (scene flags offset 0x0C).
-    /// Freestanding items, rupees, hearts, etc.
-    Collectible,
-
-    /// Gold Skulltula flags (separate from scene flags).
-    /// Stored in dedicated skulltula section at 0x0E9C.
-    GoldSkulltula,
-
-    /// Event check flags (event_chk_inf).
-    /// Global events like cutscenes watched, NPCs talked to.
-    EventChkInf,
-
-    /// Item get flags (item_get_inf).
-    /// Tracks specific item acquisitions.
-    ItemGetInf,
-
-    /// Info table flags (inf_table).
-    /// NPC conversation flags, misc game state.
-    InfTable,
-
-    /// Shop item flags.
-    /// Tracks purchased shop items.
-    Shop,
-
-    /// Scrub/merchant purchase flags.
-    /// Business scrubs and other merchants.
-    Scrub,
-
-    /// Great Fairy reward flags.
-    /// Tracks fairy fountain upgrades received.
-    GreatFairy,
-
-    /// Boss defeated flags.
-    /// Typically stored in switch or event flags.
-    Boss,
-
-    /// Song learned flags.
-    /// Stored in quest items bitfield.
-    Song,
-
-    /// Fishing pond flags.
-    /// Special handling for fishing rewards.
-    Fishing,
-
-    /// Cow/milk flags.
-    /// Playing Epona's Song to cows.
-    Cow,
-
-    /// Gossip stone flags.
-    /// Hints from gossip stones (if shuffled).
-    GossipStone,
-}
-
-impl FlagType {
-    /// Returns the byte offset within scene flags for scene-based flag types.
-    ///
-    /// Returns `None` for global flag types that aren't stored per-scene.
-    #[must_use]
-    pub const fn scene_offset(&self) -> Option<usize> {
-        match self {
-            FlagType::Chest => Some(0x00),
-            FlagType::Switch => Some(0x04),
-            FlagType::RoomClear => Some(0x08),
-            FlagType::Collectible => Some(0x0C),
-            // These are global, not per-scene
-            FlagType::GoldSkulltula
-            | FlagType::EventChkInf
-            | FlagType::ItemGetInf
-            | FlagType::InfTable
-            | FlagType::Shop
-            | FlagType::Scrub
-            | FlagType::GreatFairy
-            | FlagType::Boss
-            | FlagType::Song
-            | FlagType::Fishing
-            | FlagType::Cow
-            | FlagType::GossipStone => None,
-        }
-    }
-
-    /// Returns whether this flag type is stored per-scene.
-    #[must_use]
-    pub const fn is_scene_based(&self) -> bool {
-        self.scene_offset().is_some()
-    }
-}
-
-// ============================================================================
-// Scene ID Constants
-// ============================================================================
-
-/// OoT Scene IDs.
-///
-/// Scene IDs correspond to the index in the scene flag array.
-/// Reference: <https://wiki.cloudmodding.com/oot/Scene_Table/NTSC_1.0>
-pub mod scene {
-    // Dungeons
-    pub const DEKU_TREE: u8 = 0x00;
-    pub const DODONGOS_CAVERN: u8 = 0x01;
-    pub const JABU_JABU: u8 = 0x02;
-    pub const FOREST_TEMPLE: u8 = 0x03;
-    pub const FIRE_TEMPLE: u8 = 0x04;
-    pub const WATER_TEMPLE: u8 = 0x05;
-    pub const SPIRIT_TEMPLE: u8 = 0x06;
-    pub const SHADOW_TEMPLE: u8 = 0x07;
-    pub const BOTTOM_OF_THE_WELL: u8 = 0x08;
-    pub const ICE_CAVERN: u8 = 0x09;
-    pub const GANONS_TOWER: u8 = 0x0A;
-    pub const GERUDO_TRAINING_GROUND: u8 = 0x0B;
-    pub const THIEVES_HIDEOUT: u8 = 0x0C;
-    pub const GANONS_CASTLE: u8 = 0x0D;
-    pub const GANONS_TOWER_COLLAPSING: u8 = 0x0E;
-    pub const GANONS_CASTLE_COLLAPSING: u8 = 0x0F;
-    pub const TREASURE_CHEST_GAME: u8 = 0x10;
-
-    // Boss Rooms
-    pub const DEKU_TREE_BOSS: u8 = 0x11;
-    pub const DODONGOS_CAVERN_BOSS: u8 = 0x12;
-    pub const JABU_JABU_BOSS: u8 = 0x13;
-    pub const FOREST_TEMPLE_BOSS: u8 = 0x14;
-    pub const FIRE_TEMPLE_BOSS: u8 = 0x15;
-    pub const WATER_TEMPLE_BOSS: u8 = 0x16;
-    pub const SPIRIT_TEMPLE_BOSS: u8 = 0x17;
-    pub const SHADOW_TEMPLE_BOSS: u8 = 0x18;
-    pub const GANONDORF_BOSS: u8 = 0x19;
-    pub const GANON_BOSS: u8 = 0x1A;
-    pub const TOWER_COLLAPSE_EXTERIOR: u8 = 0x1B;
-
-    // Overworld
-    pub const MARKET_ENTRANCE_DAY: u8 = 0x1C;
-    pub const MARKET_ENTRANCE_NIGHT: u8 = 0x1D;
-    pub const MARKET_ENTRANCE_RUINS: u8 = 0x1E;
-    pub const BACK_ALLEY_DAY: u8 = 0x1F;
-    pub const BACK_ALLEY_NIGHT: u8 = 0x20;
-    pub const MARKET_DAY: u8 = 0x21;
-    pub const MARKET_NIGHT: u8 = 0x22;
-    pub const MARKET_RUINS: u8 = 0x23;
-    pub const TEMPLE_OF_TIME_EXTERIOR_DAY: u8 = 0x24;
-    pub const TEMPLE_OF_TIME_EXTERIOR_NIGHT: u8 = 0x25;
-    pub const TEMPLE_OF_TIME_EXTERIOR_RUINS: u8 = 0x26;
-    pub const KNOW_IT_ALL_BROTHERS_HOUSE: u8 = 0x27;
-    pub const MIDOS_HOUSE: u8 = 0x28;
-    pub const SARIAS_HOUSE: u8 = 0x29;
-    pub const TWINS_HOUSE: u8 = 0x2A;
-    pub const LINKS_HOUSE: u8 = 0x2B;
-    pub const KAKARIKO_HOUSE_1: u8 = 0x2C;
-    pub const BACK_ALLEY_HOUSE: u8 = 0x2D;
-    pub const BAZAAR: u8 = 0x2E;
-    pub const KOKIRI_SHOP: u8 = 0x2F;
-    pub const GORON_SHOP: u8 = 0x30;
-    pub const ZORA_SHOP: u8 = 0x31;
-    pub const KAKARIKO_POTION_SHOP: u8 = 0x32;
-    pub const MARKET_POTION_SHOP: u8 = 0x33;
-    pub const BOMBCHU_SHOP: u8 = 0x34;
-    pub const HAPPY_MASK_SHOP: u8 = 0x35;
-    pub const GERUDO_VALLEY_TENT: u8 = 0x36;
-    pub const IMPAS_HOUSE: u8 = 0x37;
-    pub const LAKESIDE_LABORATORY: u8 = 0x38;
-    pub const CARPENTERS_TENT: u8 = 0x39;
-    pub const GRAVEKEEPERS_HUT: u8 = 0x3A;
-    pub const GREAT_FAIRY_FOUNTAIN_UPGRADES: u8 = 0x3B;
-    pub const FAIRY_FOUNTAIN: u8 = 0x3C;
-    pub const GREAT_FAIRY_FOUNTAIN_SPELLS: u8 = 0x3D;
-    pub const GROTTOS: u8 = 0x3E;
-    pub const GRAVE_HEART_PIECE: u8 = 0x3F;
-    pub const GRAVE_SHIELD: u8 = 0x40;
-    pub const ROYAL_FAMILYS_TOMB: u8 = 0x41;
-    pub const SHOOTING_GALLERY: u8 = 0x42;
-    pub const TEMPLE_OF_TIME: u8 = 0x43;
-    pub const CHAMBER_OF_SAGES: u8 = 0x44;
-    pub const CASTLE_HEDGE_MAZE_DAY: u8 = 0x45;
-    pub const CASTLE_HEDGE_MAZE_NIGHT: u8 = 0x46;
-    pub const CUTSCENE_MAP: u8 = 0x47;
-    pub const WINDMILL_AND_DAMPES_GRAVE: u8 = 0x48;
-    pub const FISHING_POND: u8 = 0x49;
-    pub const CASTLE_COURTYARD: u8 = 0x4A;
-    pub const BOMBCHU_BOWLING: u8 = 0x4B;
-    pub const LON_LON_RANCH_TOWER: u8 = 0x4C;
-    pub const LON_LON_RANCH_HOUSE: u8 = 0x4D;
-    pub const GUARD_HOUSE: u8 = 0x4E;
-    pub const KAKARIKO_HOUSE_2: u8 = 0x4F;
-    pub const KAKARIKO_HOUSE_3: u8 = 0x50;
-
-    // Main Overworld Areas
-    pub const HYRULE_FIELD: u8 = 0x51;
-    pub const KAKARIKO_VILLAGE: u8 = 0x52;
-    pub const GRAVEYARD: u8 = 0x53;
-    pub const ZORA_RIVER: u8 = 0x54;
-    pub const KOKIRI_FOREST: u8 = 0x55;
-    pub const SACRED_FOREST_MEADOW: u8 = 0x56;
-    pub const LAKE_HYLIA: u8 = 0x57;
-    pub const ZORAS_DOMAIN: u8 = 0x58;
-    pub const ZORAS_FOUNTAIN: u8 = 0x59;
-    pub const GERUDO_VALLEY: u8 = 0x5A;
-    pub const LOST_WOODS: u8 = 0x5B;
-    pub const DESERT_COLOSSUS: u8 = 0x5C;
-    pub const GERUDO_FORTRESS: u8 = 0x5D;
-    pub const HAUNTED_WASTELAND: u8 = 0x5E;
-    pub const HYRULE_CASTLE: u8 = 0x5F;
-    pub const DEATH_MOUNTAIN_TRAIL: u8 = 0x60;
-    pub const DEATH_MOUNTAIN_CRATER: u8 = 0x61;
-    pub const GORON_CITY: u8 = 0x62;
-    pub const LON_LON_RANCH: u8 = 0x63;
-    pub const OUTSIDE_GANONS_CASTLE: u8 = 0x64;
-
-    /// Maximum scene ID for OoT.
-    pub const MAX_SCENE_ID: u8 = 0x64;
-
-    /// Number of scenes in OoT.
-    pub const SCENE_COUNT: usize = 101;
-}
-
-// ============================================================================
-// Flag Mapping Structure
-// ============================================================================
-
-/// Mapping from a location ID to its flag address in memory.
-///
-/// This struct represents either a complete mapping (with scene_id, flag_type,
-/// and flag_bit populated) or a stub mapping (with all optional fields as None).
-///
-/// Stub mappings are generated for all locations from world data and serve as
-/// placeholders until the actual flag addresses are researched and filled in.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FlagMapping {
-    /// The unique location identifier from OoTMM world data.
-    pub location_id: &'static str,
-
-    /// The scene ID where this flag is stored (None for unmapped stubs or global flags).
-    pub scene_id: Option<u8>,
-
-    /// The type of flag used for this location (None for unmapped stubs).
-    pub flag_type: Option<FlagType>,
-
-    /// The bit position within the flag word (None for unmapped stubs).
-    /// For 32-bit flag words, this is typically 0-31 representing a single bit,
-    /// or a full bitmask for multi-bit values.
-    pub flag_bit: Option<u32>,
-}
-
-impl FlagMapping {
-    /// Creates a new unmapped stub mapping for a location.
-    #[must_use]
-    pub const fn stub(location_id: &'static str) -> Self {
-        Self {
-            location_id,
-            scene_id: None,
-            flag_type: None,
-            flag_bit: None,
-        }
-    }
-
-    /// Creates a new fully mapped location.
-    #[must_use]
-    pub const fn mapped(
-        location_id: &'static str,
-        scene_id: u8,
-        flag_type: FlagType,
-        flag_bit: u32,
-    ) -> Self {
-        Self {
-            location_id,
-            scene_id: Some(scene_id),
-            flag_type: Some(flag_type),
-            flag_bit: Some(flag_bit),
-        }
-    }
-
-    /// Creates a new global flag mapping (no scene_id).
-    #[must_use]
-    pub const fn global(location_id: &'static str, flag_type: FlagType, flag_bit: u32) -> Self {
-        Self {
-            location_id,
-            scene_id: None,
-            flag_type: Some(flag_type),
-            flag_bit: Some(flag_bit),
-        }
-    }
-
-    /// Returns whether this is an unmapped stub.
-    #[must_use]
-    pub const fn is_stub(&self) -> bool {
-        self.flag_type.is_none()
-    }
-
-    /// Returns whether this mapping is complete (has flag type and bit).
-    #[must_use]
-    pub const fn is_mapped(&self) -> bool {
-        self.flag_type.is_some() && self.flag_bit.is_some()
-    }
-}
-
-// ============================================================================
-// Static Mapping Tables
 // ============================================================================
 
 /// All OoT location IDs extracted from embedded world data.
 ///
 /// This list is generated at compile time from the OoTMM YAML files.
-static OOT_LOCATION_IDS: Lazy<Vec<&'static str>> = Lazy::new(|| {
+pub(super) static OOT_LOCATION_IDS: Lazy<Vec<&'static str>> = Lazy::new(|| {
     let db = embedded_data::create_world_database()
         .expect("Failed to load world database for location extraction");
 
@@ -394,7 +31,7 @@ static OOT_LOCATION_IDS: Lazy<Vec<&'static str>> = Lazy::new(|| {
 });
 
 /// HashMap of location ID to FlagMapping for fast lookups.
-static OOT_MAPPINGS: Lazy<HashMap<&'static str, FlagMapping>> = Lazy::new(|| {
+pub(super) static OOT_MAPPINGS: Lazy<HashMap<&'static str, FlagMapping>> = Lazy::new(|| {
     let mut map = HashMap::new();
 
     // First, add all stub mappings from world data
@@ -3568,1648 +3205,4 @@ fn add_mapping(
         location_id,
         FlagMapping::mapped(location_id, scene_id, flag_type, flag_bit),
     );
-}
-
-// ============================================================================
-// Public API
-// ============================================================================
-
-/// Returns the flag mapping for a location ID, if it exists.
-///
-/// Returns `Some(mapping)` if the location exists (even if unmapped stub),
-/// or `None` if the location ID is not recognized.
-#[must_use]
-pub fn get_mapping(location_id: &str) -> Option<&'static FlagMapping> {
-    OOT_MAPPINGS.get(location_id)
-}
-
-/// Returns an iterator over all OoT location mappings.
-///
-/// This includes both mapped locations and unmapped stubs.
-pub fn get_all_oot_mappings() -> impl Iterator<Item = &'static FlagMapping> {
-    OOT_MAPPINGS.values()
-}
-
-/// Returns the count of all OoT locations.
-#[must_use]
-pub fn oot_location_count() -> usize {
-    OOT_MAPPINGS.len()
-}
-
-/// Returns the count of mapped (non-stub) OoT locations.
-#[must_use]
-pub fn oot_mapped_count() -> usize {
-    OOT_MAPPINGS.values().filter(|m| m.is_mapped()).count()
-}
-
-/// Returns the count of unmapped stub locations.
-#[must_use]
-pub fn oot_stub_count() -> usize {
-    OOT_MAPPINGS.values().filter(|m| m.is_stub()).count()
-}
-
-/// Returns an iterator over only the mapped (non-stub) locations.
-pub fn get_mapped_locations() -> impl Iterator<Item = &'static FlagMapping> {
-    OOT_MAPPINGS.values().filter(|m| m.is_mapped())
-}
-
-/// Returns an iterator over only the stub locations.
-pub fn get_stub_locations() -> impl Iterator<Item = &'static FlagMapping> {
-    OOT_MAPPINGS.values().filter(|m| m.is_stub())
-}
-
-/// Returns all mappings for a specific scene.
-pub fn get_mappings_for_scene(scene_id: u8) -> impl Iterator<Item = &'static FlagMapping> {
-    OOT_MAPPINGS
-        .values()
-        .filter(move |m| m.scene_id == Some(scene_id))
-}
-
-/// Returns all mappings for a specific flag type.
-pub fn get_mappings_by_flag_type(
-    flag_type: FlagType,
-) -> impl Iterator<Item = &'static FlagMapping> {
-    OOT_MAPPINGS
-        .values()
-        .filter(move |m| m.flag_type == Some(flag_type))
-}
-
-/// Returns all OoT location IDs from the world data.
-pub fn get_all_oot_location_ids() -> impl Iterator<Item = &'static str> {
-    OOT_LOCATION_IDS.iter().copied()
-}
-
-// ============================================================================
-// Master Quest Integration
-// ============================================================================
-
-// Re-export MqDungeon for convenience
-pub use ootmm::settings::MqDungeon;
-
-/// Determines if an MQ location should be active based on settings.
-///
-/// This function handles MQ locations that `MqDungeon::from_location_id` might not
-/// recognize due to non-standard naming patterns (e.g., `mq_oot_mq_ganon_pot_*`
-/// instead of `mq_oot_mq_ganon_castle_*`).
-fn is_mq_location_active(location_id: &str, settings: &RandomizerSettings) -> bool {
-    // First try the standard detection
-    if let Some(dungeon) = MqDungeon::from_location_id(location_id) {
-        return settings.is_dungeon_mq(dungeon);
-    }
-
-    // For MQ locations that from_location_id doesn't recognize,
-    // try additional pattern matching based on dungeon names in the location ID
-    // Pattern: mq_oot_mq_<dungeon_shortname>_*
-
-    // Ganon's Castle locations (handles mq_oot_mq_ganon_pot_* pattern)
-    if location_id.contains("_ganon_") {
-        return settings.is_dungeon_mq(MqDungeon::GanonsCastle);
-    }
-
-    // If we can't determine the dungeon, the MQ location should not be active
-    // with default settings (conservative approach)
-    false
-}
-
-/// Returns an iterator over active location mappings based on MQ settings.
-///
-/// This filters out locations that belong to dungeons where the MQ setting
-/// doesn't match the location type (vanilla vs MQ).
-///
-/// # Arguments
-///
-/// * `settings` - The randomizer settings containing MQ dungeon selections
-///
-/// # Example
-///
-/// ```ignore
-/// use oottracker::flag_mapping::get_active_mappings;
-/// use ootmm::settings::RandomizerSettings;
-///
-/// let settings = RandomizerSettings::default();
-/// for mapping in get_active_mappings(&settings) {
-///     println!("{}: {:?}", mapping.location_id, mapping.flag_type);
-/// }
-/// ```
-pub fn get_active_mappings(
-    settings: &RandomizerSettings,
-) -> impl Iterator<Item = &'static FlagMapping> + '_ {
-    OOT_MAPPINGS
-        .values()
-        .filter(move |m| is_location_active_for_settings(m.location_id, settings))
-}
-
-/// Returns the flag mapping for a location, considering MQ settings.
-///
-/// This function checks if the location is in an MQ-able dungeon and
-/// returns the mapping only if the location matches the current MQ setting.
-///
-/// # Arguments
-///
-/// * `location_id` - The location ID to look up
-/// * `settings` - The randomizer settings containing MQ dungeon selections
-///
-/// # Returns
-///
-/// `Some(mapping)` if the location exists and is active for current settings,
-/// `None` if the location doesn't exist or is inactive due to MQ settings.
-#[must_use]
-pub fn get_active_mapping(
-    location_id: &str,
-    settings: &RandomizerSettings,
-) -> Option<&'static FlagMapping> {
-    let mapping = OOT_MAPPINGS.get(location_id)?;
-    if is_location_active_for_settings(location_id, settings) {
-        Some(mapping)
-    } else {
-        None
-    }
-}
-
-/// Helper to check if a location is active for given settings.
-fn is_location_active_for_settings(location_id: &str, settings: &RandomizerSettings) -> bool {
-    if location_id.starts_with("mq_oot_") {
-        is_mq_location_active(location_id, settings)
-    } else {
-        settings.is_location_active(location_id)
-    }
-}
-
-/// Returns the count of active (non-MQ-filtered) locations for given settings.
-#[must_use]
-pub fn active_location_count(settings: &RandomizerSettings) -> usize {
-    OOT_MAPPINGS
-        .values()
-        .filter(|m| is_location_active_for_settings(m.location_id, settings))
-        .count()
-}
-
-/// Returns the count of active mapped (non-stub) locations for given settings.
-#[must_use]
-pub fn active_mapped_count(settings: &RandomizerSettings) -> usize {
-    OOT_MAPPINGS
-        .values()
-        .filter(|m| m.is_mapped() && is_location_active_for_settings(m.location_id, settings))
-        .count()
-}
-
-/// Returns an iterator over active mapped (non-stub) locations for given settings.
-pub fn get_active_mapped_locations(
-    settings: &RandomizerSettings,
-) -> impl Iterator<Item = &'static FlagMapping> + '_ {
-    OOT_MAPPINGS
-        .values()
-        .filter(move |m| m.is_mapped() && is_location_active_for_settings(m.location_id, settings))
-}
-
-/// Returns all mappings for a specific dungeon based on its MQ status.
-///
-/// This returns either vanilla or MQ mappings depending on the dungeon's
-/// setting in the provided settings.
-pub fn get_dungeon_mappings(
-    dungeon: MqDungeon,
-    settings: &RandomizerSettings,
-) -> impl Iterator<Item = &'static FlagMapping> + '_ {
-    let prefix = settings.get_dungeon_location_prefix(dungeon);
-    OOT_MAPPINGS
-        .values()
-        .filter(move |m| m.location_id.starts_with(prefix))
-}
-
-// ============================================================================
-// MQ Settings Conversion from Knowledge
-// ============================================================================
-
-/// Converts a Dungeon enum to the corresponding MqDungeon enum.
-///
-/// This mapping is needed to bridge between the ootr Dungeon enum used in
-/// Knowledge.mq and the ootmm MqDungeon enum used in RandomizerSettings.
-fn dungeon_to_mq_dungeon(dungeon: &Dungeon) -> MqDungeon {
-    match dungeon {
-        Dungeon::Main(MainDungeon::DekuTree) => MqDungeon::DekuTree,
-        Dungeon::Main(MainDungeon::DodongosCavern) => MqDungeon::DodongosCavern,
-        Dungeon::Main(MainDungeon::JabuJabu) => MqDungeon::JabuJabu,
-        Dungeon::Main(MainDungeon::ForestTemple) => MqDungeon::ForestTemple,
-        Dungeon::Main(MainDungeon::FireTemple) => MqDungeon::FireTemple,
-        Dungeon::Main(MainDungeon::WaterTemple) => MqDungeon::WaterTemple,
-        Dungeon::Main(MainDungeon::ShadowTemple) => MqDungeon::ShadowTemple,
-        Dungeon::Main(MainDungeon::SpiritTemple) => MqDungeon::SpiritTemple,
-        Dungeon::IceCavern => MqDungeon::IceCavern,
-        Dungeon::BottomOfTheWell => MqDungeon::BottomOfTheWell,
-        Dungeon::GerudoTrainingGround => MqDungeon::GerudoTrainingGround,
-        Dungeon::GanonsCastle => MqDungeon::GanonsCastle,
-    }
-}
-
-/// Creates RandomizerSettings from Knowledge MQ settings.
-///
-/// This function converts the `Knowledge.mq` HashMap<Dungeon, Mq> into a
-/// `RandomizerSettings` struct with the appropriate MQ dungeon selections.
-///
-/// # Arguments
-///
-/// * `mq_settings` - A HashMap mapping dungeons to their MQ status from Knowledge
-///
-/// # Returns
-///
-/// A `RandomizerSettings` with MQ dungeons configured according to the Knowledge.
-///
-/// # Example
-///
-/// ```ignore
-/// use oottracker::flag_mapping::mq_settings_from_knowledge;
-/// use oottracker::Knowledge;
-///
-/// let knowledge = Knowledge::default();
-/// let settings = mq_settings_from_knowledge(&knowledge.mq);
-/// ```
-#[must_use]
-pub fn mq_settings_from_knowledge(mq_settings: &HashMap<Dungeon, Mq>) -> RandomizerSettings {
-    let mut settings = RandomizerSettings::new();
-
-    for (dungeon, mq) in mq_settings {
-        let mq_dungeon = dungeon_to_mq_dungeon(dungeon);
-        match mq {
-            Mq::Mq => settings.set_dungeon_mq(mq_dungeon),
-            Mq::Vanilla => settings.set_dungeon_vanilla(mq_dungeon),
-        }
-    }
-
-    settings
-}
-
-// ============================================================================
-// Location Check Status
-// ============================================================================
-
-use crate::ModelState;
-
-/// Status of a location check.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub enum CheckStatus {
-    /// Location has been checked (item collected).
-    Checked,
-    /// Location has not been checked yet.
-    Unchecked,
-    /// User has decided to skip this location (won't complete it).
-    Skipped,
-    /// Check status cannot be determined (unmapped or unknown).
-    Unknown,
-}
-
-/// Accessibility status based on logic evaluation.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, serde::Serialize, serde::Deserialize)]
-pub enum Accessibility {
-    /// The location is accessible with current items and state.
-    Available,
-    /// The location is not accessible yet.
-    Unavailable,
-    /// Accessibility cannot be determined (no logic defined or evaluation failed).
-    #[default]
-    Unknown,
-}
-
-// ============================================================================
-// OoT Evaluation Context
-// ============================================================================
-
-use crate::save::{Equipment, Hookshot, MagicCapacity, Ocarina, QuestItems, Save, Upgrades};
-
-/// Context for evaluating OoT logic expressions from save data.
-///
-/// This struct provides an [`EvalContext`] implementation that reads item
-/// possession, age, and other state from OoT save data.
-pub struct OotEvalContext<'a> {
-    save: &'a Save,
-}
-
-impl<'a> OotEvalContext<'a> {
-    /// Creates a new evaluation context from save data.
-    pub fn new(save: &'a Save) -> Self {
-        Self { save }
-    }
-}
-
-impl EvalContext for OotEvalContext<'_> {
-    fn has_item(&self, item: &str, count: u32) -> bool {
-        let item_upper = item.to_uppercase();
-        let has = match item_upper.as_str() {
-            // Inventory items
-            "BOW" => self.save.inv.bow,
-            "FIRE_ARROW" | "FIRE_ARROWS" => self.save.inv.fire_arrows,
-            "ICE_ARROW" | "ICE_ARROWS" => self.save.inv.ice_arrows,
-            "LIGHT_ARROW" | "LIGHT_ARROWS" => self.save.inv.light_arrows,
-            "DINS_FIRE" | "DIN" => self.save.inv.dins_fire,
-            "FARORES_WIND" | "FARORE" => self.save.inv.farores_wind,
-            "NAYRUS_LOVE" | "NAYRU" => self.save.inv.nayrus_love,
-            "SLINGSHOT" | "FAIRY_SLINGSHOT" => self.save.inv.slingshot,
-            "OCARINA" | "OCARINA_OF_TIME" | "FAIRY_OCARINA" => {
-                self.save.inv.ocarina != Ocarina::None
-            }
-            "BOMBCHU" | "BOMBCHUS" => self.save.inv.bombchus,
-            "HOOKSHOT" => matches!(
-                self.save.inv.hookshot,
-                Hookshot::Hookshot | Hookshot::Longshot
-            ),
-            "LONGSHOT" => matches!(self.save.inv.hookshot, Hookshot::Longshot),
-            "BOOMERANG" => self.save.inv.boomerang,
-            "LENS" | "LENS_OF_TRUTH" => self.save.inv.lens,
-            "MAGIC_BEAN" | "BEANS" => self.save.inv.beans,
-            "HAMMER" | "MEGATON_HAMMER" => self.save.inv.hammer,
-
-            // Equipment
-            "KOKIRI_SWORD" => self.save.equipment.contains(Equipment::KOKIRI_SWORD),
-            "MASTER_SWORD" => self.save.equipment.contains(Equipment::MASTER_SWORD),
-            "GIANTS_KNIFE" | "BIGGORON_SWORD" | "SWORD_BIGGORON" => {
-                self.save.biggoron_sword || self.save.equipment.contains(Equipment::GIANTS_KNIFE)
-            }
-            "DEKU_SHIELD" => self.save.equipment.contains(Equipment::DEKU_SHIELD),
-            "HYLIAN_SHIELD" => self.save.equipment.contains(Equipment::HYLIAN_SHIELD),
-            "MIRROR_SHIELD" => self.save.equipment.contains(Equipment::MIRROR_SHIELD),
-            "GORON_TUNIC" => self.save.equipment.contains(Equipment::GORON_TUNIC),
-            "ZORA_TUNIC" => self.save.equipment.contains(Equipment::ZORA_TUNIC),
-            "IRON_BOOTS" => self.save.equipment.contains(Equipment::IRON_BOOTS),
-            "HOVER_BOOTS" => self.save.equipment.contains(Equipment::HOVER_BOOTS),
-
-            // Upgrades
-            "GORON_BRACELET" | "STRENGTH" => self.save.upgrades.contains(Upgrades::GORON_BRACELET),
-            "SILVER_GAUNTLETS" => self.save.upgrades.contains(Upgrades::SILVER_GAUNTLETS),
-            "GOLD_GAUNTLETS" | "GOLDEN_GAUNTLETS" => {
-                self.save.upgrades.contains(Upgrades::GOLD_GAUNTLETS)
-            }
-            "SILVER_SCALE" | "SCALE" => self.save.upgrades.contains(Upgrades::SILVER_SCALE),
-            "GOLD_SCALE" | "GOLDEN_SCALE" => self.save.upgrades.contains(Upgrades::GOLD_SCALE),
-            "BOMB_BAG" | "BOMBS" => self
-                .save
-                .upgrades
-                .intersects(Upgrades::BOMB_BAG_20 | Upgrades::BOMB_BAG_30 | Upgrades::BOMB_BAG_40),
-            "QUIVER" => self.save.upgrades.intersects(
-                Upgrades::QUIVER_50 | Upgrades::QUIVER_40 | Upgrades::from_bits_truncate(0x1),
-            ),
-            "BULLET_BAG" => self.save.upgrades.intersects(
-                Upgrades::BULLET_BAG_30 | Upgrades::BULLET_BAG_40 | Upgrades::BULLET_BAG_50,
-            ),
-            "WALLET" | "ADULTS_WALLET" | "GIANTS_WALLET" => {
-                self.save.upgrades.intersects(Upgrades::WALLET_MASK)
-            }
-
-            // Magic
-            "MAGIC" | "MAGIC_METER" => self.save.magic != MagicCapacity::None,
-            "DOUBLE_MAGIC" => self.save.magic == MagicCapacity::Large,
-
-            // Quest items - songs
-            "ZELDAS_LULLABY" | "LULLABY" => {
-                self.save.quest_items.contains(QuestItems::ZELDAS_LULLABY)
-            }
-            "EPONAS_SONG" | "EPONA" => self.save.quest_items.contains(QuestItems::EPONAS_SONG),
-            "SARIAS_SONG" | "SARIA" => self.save.quest_items.contains(QuestItems::SARIAS_SONG),
-            "SUNS_SONG" | "SUN" => self.save.quest_items.contains(QuestItems::SUNS_SONG),
-            "SONG_OF_TIME" | "TIME" => self.save.quest_items.contains(QuestItems::SONG_OF_TIME),
-            "SONG_OF_STORMS" | "STORMS" => {
-                self.save.quest_items.contains(QuestItems::SONG_OF_STORMS)
-            }
-            "MINUET_OF_FOREST" | "MINUET" => {
-                self.save.quest_items.contains(QuestItems::MINUET_OF_FOREST)
-            }
-            "BOLERO_OF_FIRE" | "BOLERO" => {
-                self.save.quest_items.contains(QuestItems::BOLERO_OF_FIRE)
-            }
-            "SERENADE_OF_WATER" | "SERENADE" => self
-                .save
-                .quest_items
-                .contains(QuestItems::SERENADE_OF_WATER),
-            "REQUIEM_OF_SPIRIT" | "REQUIEM" => self
-                .save
-                .quest_items
-                .contains(QuestItems::REQUIEM_OF_SPIRIT),
-            "NOCTURNE_OF_SHADOW" | "NOCTURNE" => self
-                .save
-                .quest_items
-                .contains(QuestItems::NOCTURNE_OF_SHADOW),
-            "PRELUDE_OF_LIGHT" | "PRELUDE" => {
-                self.save.quest_items.contains(QuestItems::PRELUDE_OF_LIGHT)
-            }
-
-            // Quest items - stones and medallions
-            "KOKIRI_EMERALD" => self.save.quest_items.contains(QuestItems::KOKIRI_EMERALD),
-            "GORON_RUBY" => self.save.quest_items.contains(QuestItems::GORON_RUBY),
-            "ZORA_SAPPHIRE" => self.save.quest_items.contains(QuestItems::ZORA_SAPPHIRE),
-            "FOREST_MEDALLION" => self.save.quest_items.contains(QuestItems::FOREST_MEDALLION),
-            "FIRE_MEDALLION" => self.save.quest_items.contains(QuestItems::FIRE_MEDALLION),
-            "WATER_MEDALLION" => self.save.quest_items.contains(QuestItems::WATER_MEDALLION),
-            "SPIRIT_MEDALLION" => self.save.quest_items.contains(QuestItems::SPIRIT_MEDALLION),
-            "SHADOW_MEDALLION" => self.save.quest_items.contains(QuestItems::SHADOW_MEDALLION),
-            "LIGHT_MEDALLION" => self.save.quest_items.contains(QuestItems::LIGHT_MEDALLION),
-
-            // Other quest items
-            "GERUDO_CARD" | "GERUDO_MEMBERSHIP_CARD" => {
-                self.save.quest_items.contains(QuestItems::GERUDO_CARD)
-            }
-            "STONE_OF_AGONY" => self.save.quest_items.contains(QuestItems::STONE_OF_AGONY),
-
-            // Bottles (check if any bottle exists)
-            "BOTTLE" | "EMPTY_BOTTLE" => self
-                .save
-                .inv
-                .bottles
-                .iter()
-                .any(|b| *b != crate::save::Bottle::None),
-
-            // Sticks/Nuts (check if capacity upgrade exists)
-            "DEKU_STICK" | "STICKS" => self
-                .save
-                .upgrades
-                .intersects(Upgrades::DEKU_STICK_CAPACITY_MASK),
-            "DEKU_NUT" | "NUTS" => self
-                .save
-                .upgrades
-                .intersects(Upgrades::DEKU_NUT_CAPACITY_MASK),
-
-            // Default - unknown item
-            _ => false,
-        };
-
-        // For count-based checks, we only support count=1 for now
-        // Most logic just checks has(ITEM) which defaults to count=1
-        has && count <= 1
-    }
-
-    fn event(&self, _name: &str) -> bool {
-        // Events require more complex tracking (event_chk_inf flags)
-        // For now, return false - events are harder to map without full flag analysis
-        false
-    }
-
-    fn setting(&self, _name: &str) -> Option<bool> {
-        // Settings are not stored in save data
-        // Return None to indicate unknown
-        None
-    }
-
-    fn setting_value(&self, _name: &str, _value: &str) -> bool {
-        // Settings are not stored in save data
-        false
-    }
-
-    fn trick(&self, _name: &str) -> bool {
-        // Tricks are user preferences, not stored in save
-        false
-    }
-
-    fn is_adult(&self) -> bool {
-        self.save.is_adult
-    }
-
-    fn is_child(&self) -> bool {
-        !self.save.is_adult
-    }
-
-    fn mm_time(&self) -> u32 {
-        // OoT doesn't have MM time
-        0
-    }
-}
-
-// ============================================================================
-// World Database Cache
-// ============================================================================
-
-/// Lazily initialized world database containing all location logic definitions.
-static WORLD_DATABASE: Lazy<Option<WorldDatabase>> =
-    Lazy::new(|| match embedded_data::create_world_database() {
-        Ok(db) => Some(db),
-        Err(e) => {
-            eprintln!(
-                "Warning: Failed to load world database for logic evaluation: {}",
-                e
-            );
-            None
-        }
-    });
-
-/// Evaluates accessibility for a location based on current save state.
-///
-/// Returns `Accessibility::Available` if the location's logic evaluates to true,
-/// `Accessibility::Unavailable` if it evaluates to false, or `Accessibility::Unknown`
-/// if the location is not found or has no logic defined.
-fn evaluate_accessibility(location_id: &str, save: &Save) -> Accessibility {
-    let Some(db) = WORLD_DATABASE.as_ref() else {
-        return Accessibility::Unknown;
-    };
-
-    let Some((location, region_id)) = db.get_location(location_id) else {
-        return Accessibility::Unknown;
-    };
-
-    // Get the region to determine which game this location belongs to
-    let Some(region) = db.get_region(region_id) else {
-        return Accessibility::Unknown;
-    };
-
-    let ctx = OotEvalContext::new(save);
-
-    // Check if the region is reachable before evaluating location logic
-    if !is_region_reachable(db, &ctx, region.game, region_id) {
-        return Accessibility::Unavailable;
-    }
-
-    let Some(logic) = &location.logic else {
-        // No logic defined means always accessible
-        return Accessibility::Available;
-    };
-
-    match eval_str(logic, &ctx) {
-        Ok(true) => Accessibility::Available,
-        Ok(false) => Accessibility::Unavailable,
-        Err(_) => Accessibility::Unknown,
-    }
-}
-
-/// Result of checking a location.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct LocationCheckResult {
-    /// The location ID.
-    pub location_id: String,
-    /// Whether the location has been checked.
-    pub status: CheckStatus,
-    /// Whether this location has a valid flag mapping.
-    pub is_mapped: bool,
-    /// Logic expression required to access this location (if available).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub logic: Option<String>,
-    /// Accessibility based on logic evaluation.
-    #[serde(default)]
-    pub accessibility: Accessibility,
-}
-
-/// Returns the logic expression for a location from the world database.
-///
-/// # Arguments
-///
-/// * `location_id` - The location ID to look up
-///
-/// # Returns
-///
-/// The logic expression as a String if found, None otherwise.
-pub fn get_location_logic(location_id: &str) -> Option<String> {
-    let db = world_database();
-    db.get_location(location_id)
-        .and_then(|(location, _)| location.logic.clone())
-}
-
-/// Checks if a specific location has been checked based on the current game state.
-///
-/// # Arguments
-///
-/// * `mapping` - The flag mapping for the location
-/// * `model` - The current model state containing game memory
-///
-/// # Returns
-///
-/// `CheckStatus::Skipped` if the user has marked this location as skipped,
-/// `CheckStatus::Checked` if the location flag is set,
-/// `CheckStatus::Unchecked` if the flag is not set,
-/// `CheckStatus::Unknown` if the location is unmapped or cannot be determined.
-#[must_use]
-pub fn check_location_status(mapping: &FlagMapping, model: &ModelState) -> CheckStatus {
-    // Check if user has marked this location as skipped
-    if model.skipped_locations.contains(mapping.location_id) {
-        return CheckStatus::Skipped;
-    }
-
-    // If the mapping is a stub (unmapped), we can't determine the status
-    if mapping.is_stub() {
-        return CheckStatus::Unknown;
-    }
-
-    let flag_type = match mapping.flag_type {
-        Some(ft) => ft,
-        None => return CheckStatus::Unknown,
-    };
-
-    let flag_bit = match mapping.flag_bit {
-        Some(fb) => fb,
-        None => return CheckStatus::Unknown,
-    };
-
-    match flag_type {
-        FlagType::Chest => {
-            if let Some(scene_id) = mapping.scene_id {
-                let scene_flags = model.ram.scene_flags();
-                let chests = scene_flags.get_chest_flags(scene_id);
-                if chests & flag_bit != 0 {
-                    CheckStatus::Checked
-                } else {
-                    CheckStatus::Unchecked
-                }
-            } else {
-                CheckStatus::Unknown
-            }
-        }
-        FlagType::Switch => {
-            if let Some(scene_id) = mapping.scene_id {
-                let scene_flags = model.ram.scene_flags();
-                let switches = scene_flags.get_switch_flags(scene_id);
-                if switches & flag_bit != 0 {
-                    CheckStatus::Checked
-                } else {
-                    CheckStatus::Unchecked
-                }
-            } else {
-                CheckStatus::Unknown
-            }
-        }
-        FlagType::RoomClear => {
-            if let Some(scene_id) = mapping.scene_id {
-                let scene_flags = model.ram.scene_flags();
-                let room_clear = scene_flags.get_room_clear_flags(scene_id);
-                if room_clear & flag_bit != 0 {
-                    CheckStatus::Checked
-                } else {
-                    CheckStatus::Unchecked
-                }
-            } else {
-                CheckStatus::Unknown
-            }
-        }
-        FlagType::Collectible => {
-            if let Some(scene_id) = mapping.scene_id {
-                let scene_flags = model.ram.scene_flags();
-                let collectible = scene_flags.get_collectible_flags(scene_id);
-                if collectible & flag_bit != 0 {
-                    CheckStatus::Checked
-                } else {
-                    CheckStatus::Unchecked
-                }
-            } else {
-                CheckStatus::Unknown
-            }
-        }
-        FlagType::GoldSkulltula => {
-            // Gold Skulltulas use a different storage mechanism
-            // Check the gold_skulltulas field in save data
-            let gs_flags = model.ram.save.gold_skulltulas.get_raw_flags();
-            if gs_flags & flag_bit != 0 {
-                CheckStatus::Checked
-            } else {
-                CheckStatus::Unchecked
-            }
-        }
-        FlagType::EventChkInf => {
-            let event_flags = model.ram.save.event_chk_inf.get_raw_flags();
-            if event_flags & flag_bit != 0 {
-                CheckStatus::Checked
-            } else {
-                CheckStatus::Unchecked
-            }
-        }
-        FlagType::ItemGetInf => {
-            let item_flags = model.ram.save.item_get_inf.get_raw_flags();
-            if item_flags & flag_bit != 0 {
-                CheckStatus::Checked
-            } else {
-                CheckStatus::Unchecked
-            }
-        }
-        FlagType::InfTable => {
-            let inf_flags = model.ram.save.inf_table.get_raw_flags();
-            if inf_flags & flag_bit != 0 {
-                CheckStatus::Checked
-            } else {
-                CheckStatus::Unchecked
-            }
-        }
-        // These flag types need special handling or are not yet implemented
-        FlagType::Shop
-        | FlagType::Scrub
-        | FlagType::GreatFairy
-        | FlagType::Boss
-        | FlagType::Song
-        | FlagType::Fishing
-        | FlagType::Cow
-        | FlagType::GossipStone => CheckStatus::Unknown,
-    }
-}
-
-/// Returns all checked locations for the current game state.
-///
-/// This function returns check status for all OoT locations. For combo randomizer
-/// tracking that includes MM locations, use `get_all_checked_locations_combo`.
-///
-/// # Arguments
-///
-/// * `model` - The current model state containing game memory
-///
-/// # Returns
-///
-/// A vector of `LocationCheckResult` for all mapped OoT locations.
-pub fn get_all_checked_locations(model: &ModelState) -> Vec<LocationCheckResult> {
-    get_mapped_locations()
-        .map(|mapping| {
-            let location_id = mapping.location_id.to_string();
-            let accessibility = evaluate_accessibility(&location_id, &model.ram.save);
-            LocationCheckResult {
-                location_id: location_id.clone(),
-                status: check_location_status(mapping, model),
-                is_mapped: mapping.is_mapped(),
-                logic: get_location_logic(&location_id),
-                accessibility,
-            }
-        })
-        .collect()
-}
-
-/// Returns all checked locations for both OoT and MM (combo randomizer).
-///
-/// This function combines OoT and MM location checks for combo randomizer tracking.
-/// MM locations are only included if `model.ram.mm_save` is Some.
-///
-/// # Arguments
-///
-/// * `model` - The current model state containing game memory
-///
-/// # Returns
-///
-/// A vector of `LocationCheckResult` for all OoT locations and MM locations (if available).
-pub fn get_all_checked_locations_combo(model: &ModelState) -> Vec<LocationCheckResult> {
-    use crate::mm_flag_mapping::{check_mm_location_status, get_mm_mapped_locations};
-
-    // Get all OoT locations
-    let mut results: Vec<LocationCheckResult> = get_mapped_locations()
-        .map(|mapping| {
-            let location_id = mapping.location_id.to_string();
-            let accessibility = evaluate_accessibility(&location_id, &model.ram.save);
-            LocationCheckResult {
-                location_id: location_id.clone(),
-                status: check_location_status(mapping, model),
-                is_mapped: mapping.is_mapped(),
-                logic: get_location_logic(&location_id),
-                accessibility,
-            }
-        })
-        .collect();
-
-    // Add MM locations if MM save data is available
-    if let Some(ref mm_save) = model.ram.mm_save {
-        let mm_results: Vec<LocationCheckResult> = get_mm_mapped_locations()
-            .map(|mapping| {
-                let location_id = mapping.location_id.to_string();
-                // MM accessibility evaluation would use MM save context
-                // For now, mark as Unknown since we don't have MM logic evaluation yet
-                LocationCheckResult {
-                    location_id: location_id.clone(),
-                    status: check_mm_location_status(mapping, mm_save),
-                    is_mapped: mapping.is_mapped(),
-                    logic: get_location_logic(&location_id),
-                    accessibility: Accessibility::Unknown,
-                }
-            })
-            .collect();
-        results.extend(mm_results);
-    }
-
-    results
-}
-
-/// Returns all checked locations as a HashMap for efficient lookup.
-///
-/// # Arguments
-///
-/// * `model` - The current model state containing game memory
-///
-/// # Returns
-///
-/// A HashMap mapping location_id to CheckStatus.
-pub fn get_checked_locations_map(model: &ModelState) -> HashMap<String, CheckStatus> {
-    get_mapped_locations()
-        .map(|mapping| {
-            (
-                mapping.location_id.to_string(),
-                check_location_status(mapping, model),
-            )
-        })
-        .collect()
-}
-
-/// Summary of checked locations.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct CheckedLocationsSummary {
-    /// Total number of mapped locations.
-    pub total_mapped: usize,
-    /// Number of checked locations.
-    pub checked_count: usize,
-    /// Number of unchecked locations.
-    pub unchecked_count: usize,
-    /// Number of skipped locations.
-    pub skipped_count: usize,
-    /// Number of locations with unknown status.
-    pub unknown_count: usize,
-    /// Number of available (accessible) unchecked locations.
-    #[serde(default)]
-    pub available_count: usize,
-    /// Number of unavailable (inaccessible) unchecked locations.
-    #[serde(default)]
-    pub unavailable_count: usize,
-    /// List of location check results.
-    pub locations: Vec<LocationCheckResult>,
-    /// Current scene ID from RAM (used for auto-scrolling the check tracker).
-    /// This is the OoT scene ID when in OoT, or the MM scene ID when in MM.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub current_scene_id: Option<u8>,
-    /// Indicates which game is currently active (for combo randomizer).
-    /// Values: "oot" for Ocarina of Time, "mm" for Majora's Mask, None if unknown.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub current_game: Option<String>,
-}
-
-/// Returns a summary of checked locations for the current game state.
-///
-/// This function includes both OoT and MM locations when MM save data is available,
-/// making it suitable for combo randomizer tracking.
-///
-/// # Arguments
-///
-/// * `model` - The current model state containing game memory
-///
-/// # Returns
-///
-/// A `CheckedLocationsSummary` containing counts and individual location statuses.
-pub fn get_checked_locations_summary(model: &ModelState) -> CheckedLocationsSummary {
-    // Use combo version to include both OoT and MM locations
-    let locations = get_all_checked_locations_combo(model);
-    let total_mapped = locations.len();
-    let checked_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Checked)
-        .count();
-    let unchecked_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Unchecked)
-        .count();
-    let skipped_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Skipped)
-        .count();
-    let unknown_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Unknown)
-        .count();
-    // Count available/unavailable only for unchecked locations
-    let available_count = locations
-        .iter()
-        .filter(|l| {
-            l.status == CheckStatus::Unchecked && l.accessibility == Accessibility::Available
-        })
-        .count();
-    let unavailable_count = locations
-        .iter()
-        .filter(|l| {
-            l.status == CheckStatus::Unchecked && l.accessibility == Accessibility::Unavailable
-        })
-        .count();
-
-    // Determine current scene and game from RAM state
-    let current_scene_id = Some(model.ram.current_scene_id);
-
-    // Determine which game is active based on MM save data presence
-    // If MM save is present and has valid data, we might be in MM
-    let current_game = if model.ram.mm_save.is_some() {
-        // In combo mode - could be either game
-        // Scene ID 0xFF typically indicates no active scene
-        if model.ram.current_scene_id == 0xFF {
-            None
-        } else {
-            // Default to OoT for now - more sophisticated detection would
-            // require checking game detection state which isn't stored in ModelState
-            Some("oot".to_string())
-        }
-    } else {
-        // No MM save means standalone OoT
-        Some("oot".to_string())
-    };
-
-    CheckedLocationsSummary {
-        total_mapped,
-        checked_count,
-        unchecked_count,
-        skipped_count,
-        unknown_count,
-        available_count,
-        unavailable_count,
-        locations,
-        current_scene_id,
-        current_game,
-    }
-}
-
-/// Returns all checked locations for the current game state, filtered by MQ settings.
-///
-/// This function filters out locations that don't match the current MQ/vanilla
-/// settings for each dungeon as stored in the model's Knowledge.
-///
-/// # Arguments
-///
-/// * `model` - The current model state containing game memory and knowledge
-///
-/// # Returns
-///
-/// A vector of `LocationCheckResult` for all active locations based on MQ settings.
-pub fn get_all_checked_locations_filtered(model: &ModelState) -> Vec<LocationCheckResult> {
-    let settings = mq_settings_from_knowledge(&model.knowledge.mq);
-    get_active_mapped_locations(&settings)
-        .map(|mapping| {
-            let location_id = mapping.location_id.to_string();
-            let accessibility = evaluate_accessibility(&location_id, &model.ram.save);
-            LocationCheckResult {
-                location_id: location_id.clone(),
-                status: check_location_status(mapping, model),
-                is_mapped: mapping.is_mapped(),
-                logic: get_location_logic(&location_id),
-                accessibility,
-            }
-        })
-        .collect()
-}
-
-/// Returns a summary of checked locations filtered by MQ settings from Knowledge.
-///
-/// This function filters out locations that don't match the current MQ/vanilla
-/// settings for each dungeon. For example, if the Deku Tree is set to vanilla
-/// in Knowledge.mq, then Master Quest Deku Tree locations will be excluded.
-///
-/// # Arguments
-///
-/// * `model` - The current model state containing game memory and knowledge
-///
-/// # Returns
-///
-/// A `CheckedLocationsSummary` containing counts and individual location statuses,
-/// filtered to only include locations matching the current MQ settings.
-///
-/// # Example
-///
-/// ```ignore
-/// use oottracker::flag_mapping::get_checked_locations_summary_filtered;
-/// use oottracker::ModelState;
-///
-/// let model = ModelState::default();
-/// let summary = get_checked_locations_summary_filtered(&model);
-/// // summary.locations will only include vanilla dungeon checks
-/// // since Knowledge.default() sets all dungeons to vanilla
-/// ```
-pub fn get_checked_locations_summary_filtered(model: &ModelState) -> CheckedLocationsSummary {
-    let locations = get_all_checked_locations_filtered(model);
-    let total_mapped = locations.len();
-    let checked_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Checked)
-        .count();
-    let unchecked_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Unchecked)
-        .count();
-    let skipped_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Skipped)
-        .count();
-    let unknown_count = locations
-        .iter()
-        .filter(|l| l.status == CheckStatus::Unknown)
-        .count();
-    // Count available/unavailable only for unchecked locations
-    let available_count = locations
-        .iter()
-        .filter(|l| {
-            l.status == CheckStatus::Unchecked && l.accessibility == Accessibility::Available
-        })
-        .count();
-    let unavailable_count = locations
-        .iter()
-        .filter(|l| {
-            l.status == CheckStatus::Unchecked && l.accessibility == Accessibility::Unavailable
-        })
-        .count();
-
-    // Determine current scene and game from RAM state
-    let current_scene_id = Some(model.ram.current_scene_id);
-
-    // Determine which game is active based on MM save data presence
-    let current_game = if model.ram.mm_save.is_some() {
-        if model.ram.current_scene_id == 0xFF {
-            None
-        } else {
-            Some("oot".to_string())
-        }
-    } else {
-        Some("oot".to_string())
-    };
-
-    CheckedLocationsSummary {
-        total_mapped,
-        checked_count,
-        unchecked_count,
-        skipped_count,
-        unknown_count,
-        available_count,
-        unavailable_count,
-        locations,
-        current_scene_id,
-        current_game,
-    }
-}
-
-// ============================================================================
-// Tests
-// ============================================================================
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_flag_type_scene_offset() {
-        assert_eq!(FlagType::Chest.scene_offset(), Some(0x00));
-        assert_eq!(FlagType::Switch.scene_offset(), Some(0x04));
-        assert_eq!(FlagType::RoomClear.scene_offset(), Some(0x08));
-        assert_eq!(FlagType::Collectible.scene_offset(), Some(0x0C));
-        assert_eq!(FlagType::GoldSkulltula.scene_offset(), None);
-        assert_eq!(FlagType::EventChkInf.scene_offset(), None);
-    }
-
-    #[test]
-    fn test_flag_type_is_scene_based() {
-        assert!(FlagType::Chest.is_scene_based());
-        assert!(FlagType::Switch.is_scene_based());
-        assert!(!FlagType::GoldSkulltula.is_scene_based());
-        assert!(!FlagType::EventChkInf.is_scene_based());
-    }
-
-    #[test]
-    fn test_flag_mapping_stub() {
-        let stub = FlagMapping::stub("test_location");
-        assert!(stub.is_stub());
-        assert!(!stub.is_mapped());
-        assert_eq!(stub.location_id, "test_location");
-        assert!(stub.scene_id.is_none());
-        assert!(stub.flag_type.is_none());
-        assert!(stub.flag_bit.is_none());
-    }
-
-    #[test]
-    fn test_flag_mapping_mapped() {
-        let mapped = FlagMapping::mapped("test_chest", scene::DEKU_TREE, FlagType::Chest, 0x01);
-        assert!(!mapped.is_stub());
-        assert!(mapped.is_mapped());
-        assert_eq!(mapped.scene_id, Some(scene::DEKU_TREE));
-        assert_eq!(mapped.flag_type, Some(FlagType::Chest));
-        assert_eq!(mapped.flag_bit, Some(0x01));
-    }
-
-    #[test]
-    fn test_flag_mapping_global() {
-        let global = FlagMapping::global("test_skulltula", FlagType::GoldSkulltula, 0x04);
-        assert!(!global.is_stub());
-        assert!(global.is_mapped());
-        assert!(global.scene_id.is_none());
-        assert_eq!(global.flag_type, Some(FlagType::GoldSkulltula));
-    }
-
-    #[test]
-    fn test_get_mapping_exists() {
-        // This should exist from the world data
-        let mappings_count = oot_location_count();
-        assert!(
-            mappings_count > 0,
-            "Should have at least some OoT locations"
-        );
-    }
-
-    #[test]
-    fn test_get_mapping_not_found() {
-        let result = get_mapping("nonexistent_location_xyz");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_location_counts() {
-        let total = oot_location_count();
-        let mapped = oot_mapped_count();
-        let stubs = oot_stub_count();
-
-        assert_eq!(total, mapped + stubs);
-        // We should have mostly stubs since only some are mapped
-        assert!(stubs > 0, "Should have some stub locations");
-    }
-
-    #[test]
-    fn test_get_mappings_for_scene() {
-        let deku_tree_mappings: Vec<_> = get_mappings_for_scene(scene::DEKU_TREE).collect();
-        // We added several Deku Tree chest mappings
-        assert!(
-            !deku_tree_mappings.is_empty(),
-            "Should have Deku Tree mappings"
-        );
-        for mapping in &deku_tree_mappings {
-            assert_eq!(mapping.scene_id, Some(scene::DEKU_TREE));
-        }
-    }
-
-    #[test]
-    fn test_get_mappings_by_flag_type() {
-        let chest_mappings: Vec<_> = get_mappings_by_flag_type(FlagType::Chest).collect();
-        assert!(!chest_mappings.is_empty(), "Should have chest mappings");
-        for mapping in &chest_mappings {
-            assert_eq!(mapping.flag_type, Some(FlagType::Chest));
-        }
-    }
-
-    #[test]
-    fn test_scene_constants() {
-        assert_eq!(scene::DEKU_TREE, 0x00);
-        assert_eq!(scene::DODONGOS_CAVERN, 0x01);
-        assert_eq!(scene::KOKIRI_FOREST, 0x55);
-        assert_eq!(scene::MAX_SCENE_ID, 0x64);
-        assert_eq!(scene::SCENE_COUNT, 101);
-    }
-
-    #[test]
-    fn test_all_location_ids_loaded() {
-        let ids: Vec<_> = get_all_oot_location_ids().collect();
-        assert!(!ids.is_empty(), "Should have loaded OoT location IDs");
-        // All IDs should start with "oot_" or "mq_oot_" (for Master Quest variants)
-        for id in &ids {
-            assert!(
-                id.starts_with("oot_") || id.starts_with("mq_oot_"),
-                "OoT location ID should start with 'oot_' or 'mq_oot_': {}",
-                id
-            );
-        }
-    }
-
-    // === Master Quest Integration Tests ===
-
-    #[test]
-    fn test_mq_active_mappings_default() {
-        use ootmm::settings::RandomizerSettings;
-
-        let settings = RandomizerSettings::default();
-
-        // With default settings (no MQ), vanilla locations should be active
-        let active: Vec<_> = get_active_mappings(&settings).collect();
-        assert!(!active.is_empty());
-
-        // All active locations should be vanilla (not MQ)
-        for mapping in &active {
-            assert!(
-                !mapping.location_id.starts_with("mq_oot_"),
-                "Default settings should not include MQ locations: {}",
-                mapping.location_id
-            );
-        }
-    }
-
-    #[test]
-    fn test_mq_active_mapping_filters() {
-        use ootmm::settings::RandomizerSettings;
-
-        let mut settings = RandomizerSettings::default();
-
-        // Vanilla Deku Tree location should be active by default
-        assert!(get_active_mapping("oot_deku_tree_compass_chest", &settings).is_some());
-
-        // MQ Deku Tree location should NOT be active by default
-        assert!(get_active_mapping("mq_oot_mq_deku_tree_compass_chest", &settings).is_none());
-
-        // Set Deku Tree to MQ
-        settings.set_dungeon_mq(MqDungeon::DekuTree);
-
-        // Now vanilla should be inactive and MQ should be active
-        assert!(get_active_mapping("oot_deku_tree_compass_chest", &settings).is_none());
-        assert!(get_active_mapping("mq_oot_mq_deku_tree_compass_chest", &settings).is_some());
-    }
-
-    #[test]
-    fn test_mq_dungeon_mappings() {
-        use ootmm::settings::RandomizerSettings;
-
-        let mut settings = RandomizerSettings::default();
-
-        // Get vanilla Deku Tree mappings
-        let vanilla_mappings: Vec<_> =
-            get_dungeon_mappings(MqDungeon::DekuTree, &settings).collect();
-        assert!(!vanilla_mappings.is_empty());
-        for mapping in &vanilla_mappings {
-            assert!(
-                mapping.location_id.starts_with("oot_deku_tree_"),
-                "Expected vanilla Deku Tree location: {}",
-                mapping.location_id
-            );
-        }
-
-        // Set to MQ and get MQ mappings
-        settings.set_dungeon_mq(MqDungeon::DekuTree);
-        let mq_mappings: Vec<_> = get_dungeon_mappings(MqDungeon::DekuTree, &settings).collect();
-        assert!(!mq_mappings.is_empty());
-        for mapping in &mq_mappings {
-            assert!(
-                mapping.location_id.starts_with("mq_oot_mq_deku_tree_"),
-                "Expected MQ Deku Tree location: {}",
-                mapping.location_id
-            );
-        }
-    }
-
-    #[test]
-    fn test_active_counts_change_with_mq() {
-        use ootmm::settings::RandomizerSettings;
-
-        let mut settings = RandomizerSettings::default();
-
-        let vanilla_count = active_location_count(&settings);
-        assert!(vanilla_count > 0);
-
-        // Set all dungeons to MQ
-        settings.set_all_dungeons_mq();
-        let mq_count = active_location_count(&settings);
-        assert!(mq_count > 0);
-
-        // Counts might differ since MQ dungeons have different check counts
-        // The important thing is that we get different locations
-        let vanilla_settings = RandomizerSettings::default();
-        let vanilla_locs: std::collections::HashSet<_> = get_active_mappings(&vanilla_settings)
-            .map(|m| m.location_id)
-            .collect();
-        let mq_locs: std::collections::HashSet<_> = get_active_mappings(&settings)
-            .map(|m| m.location_id)
-            .collect();
-
-        // The sets should be different (vanilla vs MQ dungeon locations)
-        assert_ne!(vanilla_locs, mq_locs);
-    }
-
-    // ========================================================================
-    // get_all_checked_locations_combo integration tests
-    // ========================================================================
-
-    #[test]
-    fn test_get_all_checked_locations_combo_oot_only() {
-        // Test that without MM save data, only OoT locations are returned
-        let model = ModelState::default();
-
-        // Ensure mm_save is None (it should be by default)
-        assert!(
-            model.ram.mm_save.is_none(),
-            "Default model should have no MM save data"
-        );
-
-        let locations = get_all_checked_locations_combo(&model);
-
-        // Should have OoT locations
-        assert!(!locations.is_empty(), "Should have some locations");
-
-        // All locations should be OoT (start with "oot_" or "mq_oot_" for Master Quest)
-        // No MM locations should be present
-        for loc in &locations {
-            let is_oot =
-                loc.location_id.starts_with("oot_") || loc.location_id.starts_with("mq_oot_");
-            assert!(
-                is_oot,
-                "Without MM save, all locations should be OoT: {}",
-                loc.location_id
-            );
-            assert!(
-                !loc.location_id.starts_with("mm_"),
-                "Without MM save, no MM locations should be present: {}",
-                loc.location_id
-            );
-        }
-    }
-
-    #[test]
-    fn test_get_all_checked_locations_combo_includes_mm_when_present() {
-        use crate::mm_save::MmSave;
-
-        // Create model with MM save data
-        let mut model = ModelState::default();
-        model.ram.mm_save = Some(MmSave::default());
-
-        let locations = get_all_checked_locations_combo(&model);
-
-        // Should have locations
-        assert!(!locations.is_empty(), "Should have some locations");
-
-        // Should have both OoT and MM locations
-        let has_oot = locations.iter().any(|l| l.location_id.starts_with("oot_"));
-        let has_mm = locations.iter().any(|l| l.location_id.starts_with("mm_"));
-
-        assert!(has_oot, "Should have OoT locations");
-        assert!(has_mm, "Should have MM locations when mm_save is present");
-    }
-
-    #[test]
-    fn test_get_all_checked_locations_combo_more_locations_with_mm() {
-        use crate::mm_save::MmSave;
-
-        // Get OoT-only count
-        let model_oot_only = ModelState::default();
-        let oot_only_count = get_all_checked_locations_combo(&model_oot_only).len();
-
-        // Get OoT+MM count
-        let mut model_with_mm = ModelState::default();
-        model_with_mm.ram.mm_save = Some(MmSave::default());
-        let combo_count = get_all_checked_locations_combo(&model_with_mm).len();
-
-        // Combo should have more locations than OoT-only
-        assert!(
-            combo_count > oot_only_count,
-            "Combo locations ({}) should be more than OoT-only ({})",
-            combo_count,
-            oot_only_count
-        );
-    }
-
-    #[test]
-    fn test_get_checked_locations_summary_includes_mm() {
-        use crate::mm_save::MmSave;
-
-        // Create model with MM save data
-        let mut model = ModelState::default();
-        model.ram.mm_save = Some(MmSave::default());
-
-        let summary = get_checked_locations_summary(&model);
-
-        // Summary should include both OoT and MM locations
-        let has_oot = summary
-            .locations
-            .iter()
-            .any(|l| l.location_id.starts_with("oot_"));
-        let has_mm = summary
-            .locations
-            .iter()
-            .any(|l| l.location_id.starts_with("mm_"));
-
-        assert!(has_oot, "Summary should include OoT locations");
-        assert!(
-            has_mm,
-            "Summary should include MM locations when mm_save is present"
-        );
-
-        // Total should match location count
-        assert_eq!(
-            summary.total_mapped,
-            summary.locations.len(),
-            "total_mapped should match locations count"
-        );
-
-        // Counts should add up
-        assert_eq!(
-            summary.checked_count + summary.unchecked_count + summary.unknown_count,
-            summary.total_mapped,
-            "Status counts should sum to total_mapped"
-        );
-    }
-
-    // === dungeon_to_mq_dungeon Tests ===
-
-    #[test]
-    fn test_dungeon_to_mq_dungeon_main_dungeons() {
-        // Test all main dungeon conversions
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::DekuTree)),
-            MqDungeon::DekuTree
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::DodongosCavern)),
-            MqDungeon::DodongosCavern
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::JabuJabu)),
-            MqDungeon::JabuJabu
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::ForestTemple)),
-            MqDungeon::ForestTemple
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::FireTemple)),
-            MqDungeon::FireTemple
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::WaterTemple)),
-            MqDungeon::WaterTemple
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::ShadowTemple)),
-            MqDungeon::ShadowTemple
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::Main(MainDungeon::SpiritTemple)),
-            MqDungeon::SpiritTemple
-        );
-    }
-
-    #[test]
-    fn test_dungeon_to_mq_dungeon_mini_dungeons() {
-        // Test mini dungeon conversions
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::IceCavern),
-            MqDungeon::IceCavern
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::BottomOfTheWell),
-            MqDungeon::BottomOfTheWell
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::GerudoTrainingGround),
-            MqDungeon::GerudoTrainingGround
-        );
-        assert_eq!(
-            dungeon_to_mq_dungeon(&Dungeon::GanonsCastle),
-            MqDungeon::GanonsCastle
-        );
-    }
-
-    #[test]
-    fn test_dungeon_to_mq_dungeon_exhaustive() {
-        // Ensure all Dungeon variants can be converted
-        use enum_iterator::all;
-
-        for main_dungeon in all::<MainDungeon>() {
-            let dungeon = Dungeon::Main(main_dungeon);
-            // Should not panic
-            let _ = dungeon_to_mq_dungeon(&dungeon);
-        }
-
-        // Test non-main dungeons explicitly
-        let _ = dungeon_to_mq_dungeon(&Dungeon::IceCavern);
-        let _ = dungeon_to_mq_dungeon(&Dungeon::BottomOfTheWell);
-        let _ = dungeon_to_mq_dungeon(&Dungeon::GerudoTrainingGround);
-        let _ = dungeon_to_mq_dungeon(&Dungeon::GanonsCastle);
-    }
-
-    // === mq_settings_from_knowledge Tests ===
-
-    #[test]
-    fn test_mq_settings_from_knowledge_empty() {
-        // Empty HashMap should result in all-vanilla settings
-        let mq_settings: HashMap<Dungeon, Mq> = HashMap::new();
-        let settings = mq_settings_from_knowledge(&mq_settings);
-
-        // All dungeons should be vanilla
-        assert!(!settings.is_dungeon_mq(MqDungeon::DekuTree));
-        assert!(!settings.is_dungeon_mq(MqDungeon::ForestTemple));
-        assert!(!settings.is_dungeon_mq(MqDungeon::GanonsCastle));
-    }
-
-    #[test]
-    fn test_mq_settings_from_knowledge_single_mq() {
-        let mut mq_settings: HashMap<Dungeon, Mq> = HashMap::new();
-        mq_settings.insert(Dungeon::Main(MainDungeon::ForestTemple), Mq::Mq);
-
-        let settings = mq_settings_from_knowledge(&mq_settings);
-
-        // Only Forest Temple should be MQ
-        assert!(settings.is_dungeon_mq(MqDungeon::ForestTemple));
-        assert!(!settings.is_dungeon_mq(MqDungeon::DekuTree));
-        assert!(!settings.is_dungeon_mq(MqDungeon::FireTemple));
-    }
-
-    #[test]
-    fn test_mq_settings_from_knowledge_multiple_mq() {
-        let mut mq_settings: HashMap<Dungeon, Mq> = HashMap::new();
-        mq_settings.insert(Dungeon::Main(MainDungeon::DekuTree), Mq::Mq);
-        mq_settings.insert(Dungeon::Main(MainDungeon::ForestTemple), Mq::Mq);
-        mq_settings.insert(Dungeon::GanonsCastle, Mq::Mq);
-        mq_settings.insert(Dungeon::Main(MainDungeon::WaterTemple), Mq::Vanilla);
-
-        let settings = mq_settings_from_knowledge(&mq_settings);
-
-        // Check MQ dungeons
-        assert!(settings.is_dungeon_mq(MqDungeon::DekuTree));
-        assert!(settings.is_dungeon_mq(MqDungeon::ForestTemple));
-        assert!(settings.is_dungeon_mq(MqDungeon::GanonsCastle));
-
-        // Check vanilla dungeon
-        assert!(!settings.is_dungeon_mq(MqDungeon::WaterTemple));
-
-        // Check unspecified dungeons (should be vanilla)
-        assert!(!settings.is_dungeon_mq(MqDungeon::FireTemple));
-    }
-
-    #[test]
-    fn test_mq_settings_from_knowledge_all_vanilla() {
-        let mut mq_settings: HashMap<Dungeon, Mq> = HashMap::new();
-        // Explicitly set everything to vanilla
-        mq_settings.insert(Dungeon::Main(MainDungeon::DekuTree), Mq::Vanilla);
-        mq_settings.insert(Dungeon::Main(MainDungeon::ForestTemple), Mq::Vanilla);
-        mq_settings.insert(Dungeon::GanonsCastle, Mq::Vanilla);
-
-        let settings = mq_settings_from_knowledge(&mq_settings);
-
-        // All should be vanilla
-        assert!(!settings.is_dungeon_mq(MqDungeon::DekuTree));
-        assert!(!settings.is_dungeon_mq(MqDungeon::ForestTemple));
-        assert!(!settings.is_dungeon_mq(MqDungeon::GanonsCastle));
-    }
-
-    #[test]
-    fn test_mq_settings_from_knowledge_all_mq() {
-        use enum_iterator::all;
-
-        let mut mq_settings: HashMap<Dungeon, Mq> = HashMap::new();
-
-        // Set all main dungeons to MQ
-        for main_dungeon in all::<MainDungeon>() {
-            mq_settings.insert(Dungeon::Main(main_dungeon), Mq::Mq);
-        }
-        // Set mini dungeons to MQ
-        mq_settings.insert(Dungeon::IceCavern, Mq::Mq);
-        mq_settings.insert(Dungeon::BottomOfTheWell, Mq::Mq);
-        mq_settings.insert(Dungeon::GerudoTrainingGround, Mq::Mq);
-        mq_settings.insert(Dungeon::GanonsCastle, Mq::Mq);
-
-        let settings = mq_settings_from_knowledge(&mq_settings);
-
-        // All should be MQ
-        assert!(settings.is_dungeon_mq(MqDungeon::DekuTree));
-        assert!(settings.is_dungeon_mq(MqDungeon::DodongosCavern));
-        assert!(settings.is_dungeon_mq(MqDungeon::JabuJabu));
-        assert!(settings.is_dungeon_mq(MqDungeon::ForestTemple));
-        assert!(settings.is_dungeon_mq(MqDungeon::FireTemple));
-        assert!(settings.is_dungeon_mq(MqDungeon::WaterTemple));
-        assert!(settings.is_dungeon_mq(MqDungeon::ShadowTemple));
-        assert!(settings.is_dungeon_mq(MqDungeon::SpiritTemple));
-        assert!(settings.is_dungeon_mq(MqDungeon::IceCavern));
-        assert!(settings.is_dungeon_mq(MqDungeon::BottomOfTheWell));
-        assert!(settings.is_dungeon_mq(MqDungeon::GerudoTrainingGround));
-        assert!(settings.is_dungeon_mq(MqDungeon::GanonsCastle));
-    }
-
-    #[test]
-    fn test_mq_settings_from_knowledge_integration_with_filtering() {
-        // Test that mq_settings_from_knowledge integrates properly with location filtering
-        let mut mq_settings: HashMap<Dungeon, Mq> = HashMap::new();
-        mq_settings.insert(Dungeon::Main(MainDungeon::DekuTree), Mq::Mq);
-
-        let settings = mq_settings_from_knowledge(&mq_settings);
-
-        // Vanilla Deku Tree should be inactive
-        assert!(get_active_mapping("oot_deku_tree_compass_chest", &settings).is_none());
-
-        // MQ Deku Tree should be active
-        assert!(get_active_mapping("mq_oot_mq_deku_tree_compass_chest", &settings).is_some());
-
-        // Other vanilla dungeons should still be active
-        assert!(get_active_mapping("oot_forest_temple_compass", &settings).is_some());
-    }
-
-    // === Location Count Parity Tests ===
-
-    /// Test that all OoT locations from WorldDatabase are enumerable in the flag mapping.
-    ///
-    /// This verifies that the flag_mapping module correctly loads all OoT locations
-    /// from the embedded world data without losing any. The flag_mapping module may
-    /// have additional hardcoded entries, but all YAML locations must be present.
-    #[test]
-    fn test_oot_location_count_matches_world_database() {
-        // Get OoT locations from WorldDatabase
-        let db = embedded_data::create_world_database()
-            .expect("Failed to create world database from embedded data");
-
-        let world_db_oot_count = db.locations_for_game(Game::Oot).count();
-
-        // Verify the WorldDatabase has the expected OoT location count from YAML
-        // OoT has: 31 (bosses) + 744 (dungeons) + 926 (dungeons_mq) + 1397 (overworld) = 3098
-        const EXPECTED_YAML_OOT_LOCATIONS: usize = 3098;
-        assert_eq!(
-            world_db_oot_count, EXPECTED_YAML_OOT_LOCATIONS,
-            "WorldDatabase should have {} OoT locations from YAML, got {}",
-            EXPECTED_YAML_OOT_LOCATIONS, world_db_oot_count
-        );
-
-        // Verify that every location from WorldDatabase exists in the flag mapping
-        // (i.e., the pipeline doesn't lose any locations)
-        let mut missing_locations = Vec::new();
-        for (location, _region_id) in db.locations_for_game(Game::Oot) {
-            if get_mapping(&location.id).is_none() {
-                missing_locations.push(location.id.clone());
-            }
-        }
-
-        assert!(
-            missing_locations.is_empty(),
-            "Flag mapping is missing {} OoT locations from WorldDatabase: {:?}",
-            missing_locations.len(),
-            missing_locations
-        );
-
-        // Verify the flag_mapping count is at least the WorldDatabase count
-        let flag_mapping_count = oot_location_count();
-        assert!(
-            flag_mapping_count >= world_db_oot_count,
-            "oot_location_count() ({}) should be >= WorldDatabase OoT locations ({})",
-            flag_mapping_count,
-            world_db_oot_count
-        );
-    }
 }
